Code-split the password channge page
